STM32 code transmition part was done
There was added STM32 part of the C code. Now it can receive data by
UART and store them at the FLASH memory. But there is no python client
that can send code to the STM32

Also was added test program with blinking LEDs that we can use on
demonstration